recipe-finder/backend/requirements.txt
# requirements.txt
Flask
requests
flask-cors
python-dotenv==1.0.0
pytest 
requests_mock